example request and mock
