+++ src/data/getQuestions/getQuestions.example.ts
@@ -1,0 +1,36 @@
+import { Question } from '../../types'
+
+const getQuestionsExample: Question[] = [
+  {
+    id: 1,
+    body: 'awdawd',
+    categories: [],
+    difficulty: 'e',
+    expirationDate: '2021-05-31',
+    format: 't',
+    releaseDate: '2021-04-30',
+    title: 'test'
+  },
+  {
+    id: 2,
+    body: 'awdawd',
+    categories: [],
+    difficulty: 'e',
+    expirationDate: '2021-05-31',
+    format: 't',
+    releaseDate: '2021-04-30',
+    title: 'test'
+  },
+  {
+    id: 3,
+    body: 'awdawd',
+    categories: [],
+    difficulty: 'e',
+    expirationDate: '2021-05-31',
+    format: 't',
+    releaseDate: '2021-04-30',
+    title: 'test'
+  }
+]
+
+export { getQuestionsExample }
