+++ src/data/getQuestions/index.ts
@@ -1,0 +1,2 @@
+export { getQuestions } from './getQuestions.request'
+export { getQuestionsExample } from './getQuestions.example'
