+++ src/data/index.ts
@@ -1,1 +1,1 @@
-export * from './Questions'
+export * from './getQuestions'
